WIP: infrastructure to enable and test specialized posit templates
